+++ src/test/java/org/elasticsearch/test/rest/ElasticsearchRestTestCase.java
@@ -302,86 +302,4 @@ public abstract class ElasticsearchRestTestCase extends ElasticsearchIntegration
             executableSection.execute(restTestExecutionContext);
         }
     }
-
-    // don't look any further: NO TOUCHY!
-    
-    public static class Rest0Tests extends ElasticsearchRestTestCase {
-        public Rest0Tests(@Name("yaml") RestTestCandidate testCandidate) {
-            super(testCandidate);
-        }
-        @ParametersFactory
-        public static Iterable<Object[]> parameters() throws IOException, RestTestParseException {
-            return createParameters(0, 8);
-        }
-    }
-    
-    public static class Rest1Tests extends ElasticsearchRestTestCase {
-        public Rest1Tests(@Name("yaml") RestTestCandidate testCandidate) {
-            super(testCandidate);
-        }
-        @ParametersFactory
-        public static Iterable<Object[]> parameters() throws IOException, RestTestParseException {
-            return createParameters(1, 8);
-        }
-    }
-    
-    public static class Rest2Tests extends ElasticsearchRestTestCase {
-        public Rest2Tests(@Name("yaml") RestTestCandidate testCandidate) {
-            super(testCandidate);
-        }
-        @ParametersFactory
-        public static Iterable<Object[]> parameters() throws IOException, RestTestParseException {
-            return createParameters(2, 8);
-        }
-    }
-    
-    public static class Rest3Tests extends ElasticsearchRestTestCase {
-        public Rest3Tests(@Name("yaml") RestTestCandidate testCandidate) {
-            super(testCandidate);
-        }
-        @ParametersFactory
-        public static Iterable<Object[]> parameters() throws IOException, RestTestParseException {
-            return createParameters(3, 8);
-        }
-    }
-    
-    public static class Rest4Tests extends ElasticsearchRestTestCase {
-        public Rest4Tests(@Name("yaml") RestTestCandidate testCandidate) {
-            super(testCandidate);
-        }
-        @ParametersFactory
-        public static Iterable<Object[]> parameters() throws IOException, RestTestParseException {
-            return createParameters(4, 8);
-        }
-    }
-    
-    public static class Rest5Tests extends ElasticsearchRestTestCase {
-        public Rest5Tests(@Name("yaml") RestTestCandidate testCandidate) {
-            super(testCandidate);
-        }
-        @ParametersFactory
-        public static Iterable<Object[]> parameters() throws IOException, RestTestParseException {
-            return createParameters(5, 8);
-        }
-    }
-    
-    public static class Rest6Tests extends ElasticsearchRestTestCase {
-        public Rest6Tests(@Name("yaml") RestTestCandidate testCandidate) {
-            super(testCandidate);
-        }
-        @ParametersFactory
-        public static Iterable<Object[]> parameters() throws IOException, RestTestParseException {
-            return createParameters(6, 8);
-        }
-    }
-    
-    public static class Rest7Tests extends ElasticsearchRestTestCase {
-        public Rest7Tests(@Name("yaml") RestTestCandidate testCandidate) {
-            super(testCandidate);
-        }
-        @ParametersFactory
-        public static Iterable<Object[]> parameters() throws IOException, RestTestParseException {
-            return createParameters(7, 8);
-        }
-    }
 }
